Move town center module into buldings package
